Full Address Lookup Table support, versioned message compilation, and SPL Token-2022 integration
### 1. Address Lookup Table program alignment

- Added PDA derivation helpers for lookup tables based on authority + recent slot.
- Added a `CreateLookupTable(authority, payer, recentSlot)` overload that derives the ALT address and bump internally.
- Kept explicit creation support for callers that already have the derived lookup table address.
- Added an `ExtendLookupTable` overload that supports authority-only extension flows without forcing a payer/system-program account path.
- Added ALT instruction decoding and registered the program with the shared instruction decoder.

Result: Solnet now exposes a more complete and easier-to-use ALT surface that better matches the current on-chain interface.

### 2. Versioned message and transaction support

- Added first-class version handling to versioned message serialization.
- Updated message prefix encoding from a hardcoded v0 prefix to `0x80 | version`.
- Enabled deserialization of non-zero versioned messages by preserving the parsed message version instead of rejecting anything other than v0.
- Added support for compiling v0 transactions that include address table lookups.
- Added support for compiling v1 messages and versioned transactions.
- Added null-safe address-table lookup serialization behavior.

Result: Solnet can now compile and round-trip versioned messages more accurately, including v0 messages that consume lookup tables and v1-prefixed messages.

### 3. Token-2022 public program surface

- Added a dedicated `Token2022Program` with the Token-2022 program id and native mint id.
- Added a public `Token2022ExtensionType` enum aligned to the Rust interface values.
- Implemented the core shared SPL token builders against the Token-2022 program id, including transfer, mint, burn, authority, checked, sync-native, and multisig flows.
- Added public builders for newer instruction variants that were previously missing from Solnet's shared token surface:
  - `InitializeAccount2`
  - `InitializeAccount3`
  - `InitializeMultiSignature2`
  - `InitializeMint2`
  - `GetAccountDataSize`
  - `InitializeImmutableOwner`
  - `AmountToUiAmount`
  - `UiAmountToAmount`
- Added two additional Token-2022-specific outer instructions:
  - `InitializeMintCloseAuthority`
  - `Reallocate`
- Registered Token-2022 with the shared instruction decoder.

Result: Solnet now has a usable first-pass Token-2022 integration rather than only legacy SPL Token coverage.

### 4. Legacy TokenProgram parity improvements

- Added missing legacy `TokenProgram` builders for:
  - `InitializeAccount2`
  - `InitializeAccount3`
  - `InitializeMint2`
- Extended legacy token decoding to recognize the newly added `InitializeMintCloseAuthority` and `Reallocate` instruction enum values where shared decoding is applicable.
- Added missing shared token instruction encoders to `TokenProgramData` for the newer opcode range.
- Corrected `UiAmountToAmount` payload handling to use raw UTF-8 bytes, matching the current interface behavior.

Result: Solnet's legacy token API now better matches the modern SPL token instruction set, and shared encoding/decoding behavior is reusable across both token programs.

### 5. Associated Token Account support for Token-2022

- Added overloads to derive and create associated token accounts using a caller-supplied token program id.
- Preserved existing legacy behavior by keeping the classic token program as the default path.

Result: Solnet can now derive the correct ATA for Token-2022 mints, which differs from the legacy ATA derivation path.

### 6. Documentation and test coverage

- Added a README example that shows how to create and extend an Address Lookup Table as part of a larger transaction-building workflow.
- Added focused tests for:
  - ALT derivation, extension, and decoder behavior
  - Versioned message v0/v1 serialization and deserialization
  - Versioned transaction compilation with ALT lookups
  - Token-2022 instruction building and decoding
  - Token-2022 ATA derivation and creation behavior
  - Legacy token `InitializeAccount2`, `InitializeAccount3`, and `InitializeMint2`

+++ src/Solnet.Programs/AddressLookupTableProgramInstruction.cs
@@ -19,7 +19,7 @@ namespace Solnet.Programs
         /// <summary>
         /// Represents the instruction types for the <see cref="AddressLookupTableProgram"/>.
         /// </summary>
-        internal enum Values : byte
+        internal enum Values : uint
         {
             CreateLookupTable = 0,
             FreezeLookupTable = 1,
